basic functionality for supporting group experiments
