+++ pyensemble/models.py
@@ -67,8 +67,8 @@ class Question(models.Model):
 
     forms = models.ManyToManyField('Form', through='FormXQuestion')
 
-    class Meta:
-        unique_together = (("_unique_hash", "data_format"),)
+    # class Meta:
+    #     unique_together = (("_unique_hash", "data_format"),)
 
     def __unicode__(self):
         return self.text
@@ -82,7 +82,7 @@ class Question(models.Model):
         if self.text:
             m = hashlib.md5()
             m.update(self.text.encode('utf-8'))
-            self._unique_hash = m.digest()
+            self._unique_hash = m.hexdigest()
         else:
             self._unique_hash = ''
 
@@ -378,6 +378,7 @@ class ExperimentXForm(models.Model):
         ('form_consent','form_consent'),
         ('form_subject_register','form_subject_register'),
         ('form_subject_email','form_subject_email'),
+        ('group_trial','group_trial'),
     ]
 
     form_handler = models.CharField(max_length=50, blank=True, choices=FORM_HANDLER_OPTIONS, default='form_generic')
@@ -606,10 +607,16 @@ class GroupSubject(models.Model):
 class GroupSession(models.Model):
     group = models.ForeignKey('Group', db_constraint=True, on_delete=models.CASCADE)
     experiment = models.ForeignKey('Experiment', db_constraint=True, on_delete=models.CASCADE)
-    ticket = models.ForeignKey('Ticket', db_constraint=True, on_delete=models.CASCADE, related_name='+')
+    ticket = models.OneToOneField('Ticket', db_constraint=True, on_delete=models.CASCADE)
     start_datetime = models.DateTimeField(blank=True, null=True, auto_now_add=True)
     end_datetime = models.DateTimeField(blank=True, null=True)
     experimenter_attached = models.BooleanField(default=False)
+
+    # Mechanism for saving overall session parameters
+    params = models.JSONField(default=dict)
+
+    # Mechanism for caching current context
+    context = models.JSONField(default=dict)
 
     # Mechanism for indicating session state
     class States(models.IntegerChoices):
@@ -624,6 +631,10 @@ class GroupSession(models.Model):
     terminal_states = [States.COMPLETED, States.ABORTED]
 
     @property
+    def num_users(self):
+        return self.user_session_set.count()
+
+    @property
     def modifiable(self):
         self._modifiable=True
         if self.state in terminal_states:
@@ -631,8 +642,29 @@ class GroupSession(models.Model):
 
         return self._modifiable
 
-    # Mechanism for caching current context
-    context = models.JSONField(null=True)
+    def responding_complete(self, trial_num):
+        self._responding_complete = False
+        if trial_num == 0:
+            self._responding_complete =  True
+
+        else:
+            # Search the trial_info field the Response table contains an entry for the designated trial for all session users
+            responded = Response.objects.filter(session__in=self.user_session_set, trial_info__trial_num=trial_num)
+
+            if responded.count() == self.num_users:
+                self._responding_complete =  True
+
+        return self._responding_complete
+
+    def users_ready(self):
+        # Get number of users attached to this session
+        num_users = self.groupsessionsubjectsession_set.count()
+
+        # Get number of users in ready state
+        ready_users = self.groupsessionsubjectsession_set.filter(state=GroupSessionSubjectSession.States.READY)
+
+        return num_users == ready_users.count()
+
 
     class Meta:
         unique_together = (("group","experiment","ticket"),)
@@ -647,6 +679,14 @@ class GroupSession(models.Model):
 class GroupSessionSubjectSession(models.Model):
     group_session = models.ForeignKey('GroupSession', db_constraint=True, on_delete=models.CASCADE)
     user_session = models.ForeignKey('Session', db_constraint=True, on_delete=models.CASCADE)
+
+    # Mechanism for indicating participant session state
+    class States(models.IntegerChoices):
+        UNKNOWN = 0
+        READY = 1
+        RESPONSE_PENDING = 2
+    
+    state = models.PositiveSmallIntegerField(choices=States.choices, default=States.UNKNOWN)
 
     class Meta:
         unique_together = (('group_session','user_session'),)
