+++ pyensemble/templates/pyensemble/handlers/form_generic.html
@@ -1,5 +1,4 @@
 {% extends "pyensemble/presentation_base.html" %}
-{% load static %}
 
 {% load crispy_forms_tags %}
