+++ requirements.txt
@@ -1,5 +1,7 @@
+appnope==0.1.0
 asgiref==3.2.10
 backcall==0.1.0
+bokeh==2.3.2
 certifi==2020.6.20
 cffi==1.13.2
 chardet==3.0.4
@@ -18,16 +20,21 @@ idna==2.10
 ipython==7.11.1
 ipython-genutils==0.2.0
 jedi==0.15.2
+Jinja2==3.0.1
 kiwisolver==1.2.0
+MarkupSafe==2.0.1
 matplotlib==3.2.1
 mod-wsgi==4.7.1
 mysqlclient==1.4.6
 numpy==1.18.2
 oauthlib==3.1.0
+packaging==20.9
 pandas==1.0.4
 parso==0.5.2
 pexpect==4.7.0
 pickleshare==0.7.5
+Pillow==8.2.0
+polling2==0.5.0
 prompt-toolkit==3.0.2
 ptyprocess==0.6.0
 pyasn1==0.4.8
@@ -42,6 +49,7 @@ python-Levenshtein==0.12.1
 python-memcached==1.59
 python3-openid==3.2.0
 pytz==2019.3
+PyYAML==5.4.1
 requests==2.25.1
 requests-oauthlib==1.3.0
 rsa==4.7
@@ -51,6 +59,8 @@ six==1.13.0
 social-auth-app-django==3.4.0
 social-auth-core==3.3.3
 sqlparse==0.3.0
+tornado==6.1
 traitlets==4.3.3
+typing-extensions==3.10.0.0
 urllib3==1.26.5
 wcwidth==0.1.8
